ref(auto_config): Split FrameInfo into its own module
The main purpose for this change is to separate the frame info logic from the code mappings module.
