+++ src/sentry/issues/auto_source_code_config/errors.py
@@ -1,0 +1,18 @@
+class UnexpectedPathException(Exception):
+    pass
+
+
+class UnsupportedFrameInfo(Exception):
+    pass
+
+
+class NeedsExtension(Exception):
+    pass
+
+
+class MissingModuleOrAbsPath(Exception):
+    pass
+
+
+class DoesNotFollowJavaPackageNamingConvention(Exception):
+    pass
