+++ src/sentry/issues/auto_source_code_config/frame_info.py
@@ -1,0 +1,128 @@
+from collections.abc import Mapping
+from typing import Any
+
+from sentry.integrations.source_code_management.repo_trees import get_extension
+
+from .constants import SECOND_LEVEL_TLDS, STACK_ROOT_MAX_LEVEL
+from .errors import (
+    DoesNotFollowJavaPackageNamingConvention,
+    MissingModuleOrAbsPath,
+    NeedsExtension,
+    UnsupportedFrameInfo,
+)
+from .utils.misc import get_straight_path_prefix_end_index
+from .utils.platform import PlatformConfig
+
+
+class FrameInfo:
+    def __init__(self, frame: Mapping[str, Any], platform: str | None = None) -> None:
+        if platform:
+            platform_config = PlatformConfig(platform)
+            if platform_config.extracts_filename_from_module():
+                self.frame_info_from_module(frame)
+                return
+
+        frame_file_path = frame["filename"]
+        frame_file_path = self.transformations(frame_file_path)
+
+        # Using regexes would be better but this is easier to understand
+        if (
+            not frame_file_path
+            or frame_file_path[0] in ["[", "<"]
+            or frame_file_path.find(" ") > -1
+            or frame_file_path.find("/") == -1
+        ):
+            raise UnsupportedFrameInfo("This path is not supported.")
+
+        if not get_extension(frame_file_path):
+            raise NeedsExtension("It needs an extension.")
+
+        start_at_index = get_straight_path_prefix_end_index(frame_file_path)
+
+        # We normalize the path to be as close to what the path would
+        # look like in the source code repository, hence why we remove
+        # the straight path prefix and drive letter
+        self.normalized_path = frame_file_path[start_at_index:]
+        if start_at_index == 0:
+            self.stack_root = frame_file_path.split("/")[0]
+        else:
+            slash_index = frame_file_path.find("/", start_at_index)
+            self.stack_root = frame_file_path[0:slash_index]
+
+    def transformations(self, frame_file_path: str) -> str:
+        self.raw_path = frame_file_path
+
+        is_windows_path = False
+        if "\\" in frame_file_path:
+            is_windows_path = True
+            frame_file_path = frame_file_path.replace("\\", "/")
+
+        # Remove leading slash if it exists
+        if frame_file_path[0] == "/" or frame_file_path[0] == "\\":
+            frame_file_path = frame_file_path[1:]
+
+        # Remove drive letter if it exists
+        if is_windows_path and frame_file_path[1] == ":":
+            frame_file_path = frame_file_path[2:]
+            # windows drive letters can be like C:\ or C:
+            # so we need to remove the slash if it exists
+            if frame_file_path[0] == "/":
+                frame_file_path = frame_file_path[1:]
+
+        return frame_file_path
+
+    def frame_info_from_module(self, frame: Mapping[str, Any]) -> None:
+        if frame.get("module") and frame.get("abs_path"):
+            stack_root, filepath = get_path_from_module(frame["module"], frame["abs_path"])
+            self.stack_root = stack_root
+            self.raw_path = filepath
+            self.normalized_path = filepath
+        else:
+            raise MissingModuleOrAbsPath("Investigate why the data is missing.")
+
+    def __repr__(self) -> str:
+        return f"FrameInfo: {self.raw_path}"
+
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, FrameInfo):
+            return False
+        return self.raw_path == other.raw_path
+
+
+# Based on # https://github.com/getsentry/symbolicator/blob/450f1d6a8c346405454505ed9ca87e08a6ff34b7/crates/symbolicator-proguard/src/symbolication.rs#L450-L485
+def get_path_from_module(module: str, abs_path: str) -> tuple[str, str]:
+    """This attempts to generate a modified module and a real path from a Java module name and filename.
+    Returns a tuple of (stack_root, source_path).
+    """
+    # An `abs_path` is valid if it contains a `.` and doesn't contain a `$`.
+    if "$" in abs_path or "." not in abs_path:
+        # Split the module at the first '$' character and take the part before it
+        # If there's no '$', use the entire module
+        file_path = module.split("$", 1)[0] if "$" in module else module
+        stack_root = module.rsplit(".", 1)[0].replace(".", "/") + "/"
+        return stack_root, file_path.replace(".", "/")
+
+    if "." not in module:
+        raise DoesNotFollowJavaPackageNamingConvention
+
+    # Gets rid of the class name
+    parts = module.rsplit(".", 1)[0].split(".")
+    dirpath = "/".join(parts)
+    # a.Bar, Bar.kt -> stack_root: a/, file_path:  a/Bar.kt
+    granularity = 1
+
+    if len(parts) > 1:
+        # com.example.foo.bar.Baz$InnerClass, Baz.kt ->
+        #    stack_root: com/example/foo/
+        #    file_path:  com/example/foo/bar/Baz.kt
+        granularity = STACK_ROOT_MAX_LEVEL - 1
+
+        if parts[1] in SECOND_LEVEL_TLDS:
+            # uk.co.example.foo.bar.Baz$InnerClass, Baz.kt ->
+            #    stack_root: uk/co/example/foo/
+            #    file_path:  uk/co/example/foo/bar/Baz.kt
+            granularity = STACK_ROOT_MAX_LEVEL
+
+    stack_root = "/".join(parts[:granularity]) + "/"
+    file_path = f"{dirpath}/{abs_path}"
+    return stack_root, file_path
